pyproject.toml
[project]
name = "frcnet"
version = "0.1.0"
description = "FRCNet: resolution-content disentangled explicit-unknown network"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.8",
    "PyYAML>=6.0",
    "pytest>=8.0",
    "scikit-learn>=1.5",
    "torch>=2.3",
    "torchvision>=0.18",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]
